model/requirements.txt
transformers
torch
fastapi
re
llama-cpp-python
asyncio
nltk
string